+++ src/plugin.hpp
@@ -7,6 +7,8 @@ using namespace rack;
 // Declare the Plugin, defined in plugin.cpp
 extern Plugin* pluginInstance;
 
+#include "Controls.hpp"
+
 // Declare each Model, defined in each module source file
 // extern Model* modelMyModule;
 extern Model* modelBtoggler;
@@ -31,3 +33,7 @@ extern Model* modelDrummer4Plus;
 extern Model* modelSickoSampler;
 extern Model* modelDrumPlayerXtra;
 extern Model* modelWavetabler;
+extern Model* modelCVswitcher;
+extern Model* modelCVrouter;
+extern Model* modelAdder8;
+extern Model* modelClocker;